Moved repeated functions to "Common"
