+++ TestClientStream/Program.cs
@@ -202,7 +202,7 @@
             }
             catch (Exception e)
             {
-                LogException(e);
+                Common.LogException("StreamReceived", e);
                 return false;
             }
         }
@@ -274,18 +274,6 @@
         {
             Console.WriteLine("Server disconnected");
             return true;
-        }
-
-        private static void LogException(Exception e)
-        {
-            Console.WriteLine("================================================================================");
-            Console.WriteLine(" = Exception Type: " + e.GetType().ToString());
-            Console.WriteLine(" = Exception Data: " + e.Data);
-            Console.WriteLine(" = Inner Exception: " + e.InnerException);
-            Console.WriteLine(" = Exception Message: " + e.Message);
-            Console.WriteLine(" = Exception Source: " + e.Source);
-            Console.WriteLine(" = Exception StackTrace: " + e.StackTrace);
-            Console.WriteLine("================================================================================");
         }
     }
 }
