+++ TestMultiClient/Program.cs
@@ -1,7 +1,6 @@
 ﻿namespace TestMultiClient
 {
     using System;
-    using System.Security.Cryptography;
     using System.Threading;
     using System.Threading.Tasks;
     using WatsonTcp;
@@ -22,8 +21,8 @@
         private static void Main()
         {
             rng = new Random((int)DateTime.Now.Ticks);
-            data = InitByteArray(65536, 0x00);
-            Console.WriteLine("Data MD5: " + BytesToHex(Md5(data)));
+            data = Common.InitByteArray(65536, 0x00);
+            Console.WriteLine("Data MD5: " + Common.BytesToHex(Common.Md5(data)));
 
             Console.WriteLine("Starting server");
             server = new WatsonTcpServer(null, serverPort)
@@ -114,45 +113,6 @@
         {
             // Console.WriteLine("[server] msg from server: " + BytesToHex(Md5(data)) + " (" + data.Length + " bytes)");
             return true;
-        }
-
-        public static byte[] InitByteArray(int count, byte val)
-        {
-            byte[] ret = new byte[count];
-            for (int i = 0; i < ret.Length; i++)
-            {
-                ret[i] = val;
-            }
-
-            return ret;
-        }
-
-        private static byte[] Md5(byte[] data)
-        {
-            if (data == null || data.Length < 1)
-            {
-                return null;
-            }
-
-            using (MD5 m = MD5.Create())
-            {
-                return m.ComputeHash(data);
-            }
-        }
-
-        public static string BytesToHex(byte[] bytes)
-        {
-            if (bytes == null)
-            {
-                return null;
-            }
-
-            if (bytes.Length < 1)
-            {
-                return null;
-            }
-
-            return BitConverter.ToString(bytes).Replace("-", String.Empty);
         }
     }
 }
